Refactors to a new file the db fixture.
